% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMetaAssociation.R
\name{bulkSumlog}
\alias{bulkSumlog}
\title{Combine p-values by the sum of logs method (bulk)}
\usage{
bulkSumlog(m)
}
\arguments{
\item{m}{A matrix of p-values (0 < p <= 1)}
}
\value{
A float vector of meta p-values corresponding to each row of the
initial matrix.
}
\description{
Combine p-values by the sum of logs method, also known as
Fisher's method
}
\details{
Implementation of \code{\link[metap]{sumlog}} for matrices. Combines
p-values in rows of the matrix.
}
\examples{
data <- matrix(runif(100), 10, 10)
bulkSumlog(data)
}
