% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMetaAssociation.R
\name{calcMetaAssociation}
\alias{calcMetaAssociation}
\title{Calculate Meta P-Values Of Association}
\usage{
calcMetaAssociation(assocTable, adjust = "fdr", threshold = 1)
}
\arguments{
\item{assocTable}{a double matrix of p-values, result of
\code{calculateMassContingencyTablePvalues}}

\item{adjust}{multiple testing correction method
(see \code{\link{p.adjust.methods}})}

\item{threshold}{cutoff value for adjusted p-value}
}
\value{
A \code{data.frame} with class \code{'MetaAssociationTable'}. It has
the following columns:
\describe{
\item{Meta.P.Value}{raw meta p-values}
\item{Adj.Meta.P.Value}{meta p-values adjusted with one of the multiple
testing correction methods}
}

It also has the following attributes
\describe{
\item{adjustMethod}{method used for multiple testing correciton}
\item{experimentCount}{number of experiments used in analysis}
}
}
\description{
Calculate meta p-values of association for every motif across
all experiments with Fisher's method
}
\examples{
elements <- 10
experiments <- 10
elemNames <- paste0('elem', 1:elements)
expNames <- paste0('exp', 1:experiments)
data <- matrix(runif(elements*experiments), elements, experiments,
               dimnames=list(elemNames, expNames))
calcMetaAssociation(data, 'fdr')
}
