% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\alias{MA_AT_auxin}
\alias{AT_AGI}
\title{Example Datasets}
\format{An object of class \code{list} of length 10.}
\usage{
MA_AT_auxin

AT_AGI
}
\description{
This is the example data that is used in vignette:
\code{vignette("quickstart", package = "metaRE")}

\describe{
    \item{MA_AT_auxin}{List of NCBI GEO microarray datasets for auxin
    treatment of \emph{A. thaliana}}
    \item{AT_AGI}{List of AGI identifiers of genes used in analysis}
}
}
\keyword{datasets}
