% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClassifcation.R
\name{GeneClassificationMatrix}
\alias{GeneClassificationMatrix}
\alias{geneCounts}
\title{Gene Classification With Dense Structure}
\usage{
GeneClassificationMatrix(mat)

geneCounts(x)
}
\arguments{
\item{mat}{a logical matrix with rows corresponding to genes and columns
corresponding to experiments.}

\item{x}{a \code{GeneClassificationMatrix} object}
}
\value{
\code{GeneClassificationMatrix} returns new \code{GeneClassificationMatrix}
object


\code{geneCounts} returns geneCounts from \code{GeneClassificationMatrix}
}
\description{
Wrap a logical matrix which describes which genes were
differentially expressed in the give experiments for use in
\code{\link{calculateMassContingencyTablePvalues}}
}
\details{
\code{GeneClassificationMatrix} adds \code{geneCounts} attribute to
\code{mat} which describes column sums of the matrix and adds
\code{'GeneClassificationMatrix'} to \code{mat} classes.
}
\examples{
data <- matrix(runif(30) < 0.5, 6, 5)
rownames(data) <- paste0('gene', 1:6)
colnames(data) <- paste0('exp', 1:5)
gcm <- GeneClassificationMatrix(data)

gcm
geneCounts(gcm)
}
