% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGEO.R
\name{prepareGEO}
\alias{prepareGEO}
\title{Prepare GEO Samples for differential expression analysis.}
\usage{
prepareGEO(control, treatment, isLog2, id_col = "ID_REF",
  data_col = "VALUE")
}
\arguments{
\item{control}{GSM IDs of control samples.}

\item{treatment}{GSM IDs of treatment samples.}

\item{isLog2}{logical flag which indicates whether the original data is in
log2 scale.}

\item{id_col}{column name for row ids}

\item{data_col}{column name for expression values}
}
\value{
a data frame with a column for each sample with log2 expression
levels for each gene.
}
\description{
Download given control and treatment samples from NCBI GEO
database, convert to log2 scale if necessary.
}
\examples{
prepareGEO(
    c("GSM469656", "GSM469657", "GSM469658"),
    c("GSM469650", "GSM469651", "GSM469652"),
    FALSE
)
}
