% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMetaAssociation.R
\name{testRegulationHypotheses}
\alias{testRegulationHypotheses}
\title{Test hypotheses for association with gene regulation}
\usage{
testRegulationHypotheses(hypothesesClasses, annotationClasses,
  alternative = c("greater", "less", "two.sided"), adjust = "fdr",
  threshold = 1)
}
\arguments{
\item{hypothesesClasses}{An object of \code{\link{GeneClassificationSparse}}
class, usually describes sets of genes in which motifs are present}

\item{annotationClasses}{An object of \code{\link{GeneClassificationMatrix}}
class, usually describes sets of genes which were differentially expressed in
experiments}

\item{alternative}{indicates the alternative hypothesis and must be one of
"two.sided", "greater" or "less"}

\item{adjust}{multiple testing correction method}

\item{threshold}{cutoff value for adjusted p-value
(see \code{\link{p.adjust.methods}})}
}
\value{
Same as \code{\link{calcMetaAssociation}}.
}
\description{
Test given set of hypotheses for association with gene
regulation performing meta-analysis over many experiments.
}
\details{
This is basically a shorthand for
\code{\link{calculateMassContingencyTablePvalues}} followed by
\code{\link{calcMetaAssociation}}
}
\examples{
elements <- 5
genes <- 200
experiments <- 5

geneNames <- paste0('gene', 1:genes)
expNames <- paste0('exp', 1:experiments)
elemNames <- paste0('elem', as.character(1:elements))

gcm <- GeneClassificationMatrix(
    matrix(runif(genes*experiments)<0.05, genes, experiments,
    dimnames=list(geneNames, expNames))
)

gcs <- GeneClassificationSparse(
    setNames(lapply(1:elements, function(x) {
        sample(1:genes, as.integer(runif(1, max=genes)))
    }), elemNames),
    geneNames
)

test <- testRegulationHypotheses(gcs, gcm)
print(test)
}
\seealso{
\code{\link{calculateMassContingencyTablePvalues}},
\code{\link{calcMetaAssociation}}
}
