% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerateMotifs.R
\name{enumerateRepeats}
\alias{enumerateRepeats}
\title{Enumerate Repeats}
\usage{
enumerateRepeats(regulatoryRegions, k, minSpacer, maxSpacer, rc = TRUE,
  output = c("genes", "counts", "positions"))
}
\arguments{
\item{regulatoryRegions}{named charachter vector of nucleotide strings}

\item{k}{size of kmers}

\item{minSpacer}{minimal distance in base pairs between core and a second
motif}

\item{maxSpacer}{maximal distance in base pairs between core and a second
motif}

\item{rc}{boolean, \code{TRUE} if motifs should be considered as equal to
their reverse complements (default \code{rc=TRUE})}

\item{output}{in which format the data should be returned
(see \code{\link{enumerateMotifs}})}

\item{fuzzySpacer}{if \code{TRUE}, dyads with the same core and partner
oligomer but different spacers will be counted as the same dyad.}

\item{fuzzyOrder}{if \code{TRUE}, dyads with the same core, partner oligomer
and spacer but different order (core is left/right part of dyad) will be
counted as the same dyad.}

\item{fuzzyOrientation}{if \code{TRUE} and \code{rc=TRUE} then dyads with
partner and its reverse complement will be counted as the same dyad.}
}
\description{
Given a list of named regulatory regions, enumerate all possible
repeats with the defined spacer range and return data on their positions in
these regions.
}
\examples{
test_sequences <- c(
    gene1='acgtacgtacgt'
)
k <- 4
enumerateRepeats(test_sequences, k, -2, 4, output='positions')
}
\seealso{
\code{\link{enumerateMotifs}}, \code{\link{enumerateDyadsWithCore}}
}
