% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaWrappers.R
\name{preprocessDEGs}
\alias{preprocessDEGs}
\alias{processMicroarray}
\alias{processRNACounts}
\alias{preprocessGeneExpressionData}
\title{Extract DEGs from microarrays and RNA-Seq}
\usage{
processMicroarray(df, treatment, control, classes, adjust = "none")

processRNACounts(df, treatment, control, classes, adjust = "none")

preprocessGeneExpressionData(dataList, classes, adjust = "none")
}
\arguments{
\item{df}{data frame where column names are samples, row names are gene
names, data is either expression levels or rna counts of genes in given samples}

\item{treatment}{character vector, names of treamtment samples}

\item{control}{character vector, names of control samples}

\item{classes}{named list of functions, each function must take one argument
- a data frame which is returned by \code{\link[limma]{topTable}} function from
\code{\link{limma}} package, and must return a logical vector - genes that
are considered as DEGs.}

\item{adjust}{multiple correction method (see \code{\link{p.adjust}})}

\item{dataList}{named list of experiments where each item is a list with the
following items:
\describe{
    \item{data}{same as \code{df} parameter}
    \item{treatment}{same as \code{treament} parameter}
    \item{control}{same as \code{control} parameter}
    \item{type}{either 'MA' or 'RNA', describes the type of the dataset}
}}
}
\value{
\code{processMicroarray} and \code{processRNACounts} return data
frame with a column for each class of DEGs as defined in \code{classes}
parameter. Each column is a logical vector which tells if a gene (row)
belongs to this DEG class.

\code{preprocessGeneExpressionData} returns a named list of
\code{\link{GeneClassifcationMatrix}}, each matrix corresponds to an item in
\code{classes}
}
\description{
Given microarray or RNA-Seq data extract DEGs given user-defined
thresholds on p-values and fold changes.
}
\details{
The search is perfromed using \code{\link{limma}} and
\code{\link{edgeR}} packages. User should define the thresholds on DEGs using
\code{classes} parameters. The following columns as returned by
\code{\link[limma]{topTable}} are suggested to be used for filtering:
\describe{
    \item{logFC}{log2-fold-change}
    \item{P.Value}{raw p-value}
    \item{adj.P.Value}{adjusted p-value with \code{adjust} method}
}
}
\examples{
# The following is taken from limma::lmFit examples
# Simulate gene expression data for 100 probes and 6 microarrays
# Microarray are in two groups
# First two probes are differentially expressed in second group
# Std deviations vary between genes with prior df=4
sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- matrix(rnorm(100*6,sd=sd),100,6)
rownames(y) <- paste("Gene",1:100)
colnames(y) <- paste("Experiment", 1:6)
control <- colnames(y)[1:3]
treatment <- colnames(y)[4:6]
y[1:2,4:6] <- y[1:2,4:6] + 2
ma_data <- y

rna_data <- matrix(as.integer(exp(y)*100), 100, 6)
rownames(rna_data) <- paste("Gene",1:100)
colnames(rna_data) <- paste("Experiment", 1:6)

dataList <- list(
   exp1=list(data=ma_data, control=control, treatment=treatment, type='MA'),
   exp2=list(data=rna_data, control=control, treatment=treatment, type='RNA')
)

classes <- list(
    up=function(df) df$logFC > 0 & df$adj.P.Val < 0.1,
    down=function(df) df$logFC < 0 & df$adj.P.Val < 0.1
)

processMicroarray(ma_data, treatment, control, classes, 'fdr')
processRNACounts(rna_data, treatment, control, classes, 'fdr')

preprocessGeneExpressionData(dataList, classes, 'fdr')

}
